Refactored replacements UI into Replacements class
